web: wsgi: Implement the role-based access model for collaborators.
* src/djehuty/web/wsgi.py: Implement permission checks for
  collaborators throughout the API and UI calls.

+++ src/djehuty/web/wsgi.py
@@ -1791,6 +1791,10 @@ class ApiServer:
             if dataset is None:
                 return self.error_403 (request)
 
+            if not (not value_or (dataset, "is_shared_with_me", False) or
+                    value_or (dataset, "metadata_read", False)):
+                return self.error_403 (request)
+
             # Pre-Djehuty datasets may not have a Git UUID. We therefore
             # assign one when needed.
             if "git_uuid" not in dataset:
@@ -1838,6 +1842,10 @@ class ApiServer:
                                                    is_published=False)
 
             if dataset is None:
+                return self.error_403 (request)
+
+            if not (not value_or (dataset, "is_shared_with_me", False) or
+                    value_or (dataset, "metadata_remove", False)):
                 return self.error_403 (request)
 
             container_uuid = dataset["container_uuid"]
@@ -3804,6 +3812,10 @@ class ApiServer:
         try:
             dataset         = self.__dataset_by_id_or_uri (dataset_id, account_uuid=None, is_latest=True)
 
+            if not (not value_or (dataset, "is_shared_with_me", False) or
+                    value_or (dataset, "metadata_read", False)):
+                return self.error_403 (request)
+
             # Passing along the base_url here to generate the API links.
             dataset["base_url"] = self.base_url
 
@@ -4067,6 +4079,10 @@ class ApiServer:
                 if dataset_uri is None:
                     return self.response ("[]")
 
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_read", False)):
+                    return self.error_403 (request)
+
                 dataset["doi"]  = self.__standard_doi (dataset["container_uuid"],
                                                        version = None,
                                                        container_doi = value_or_none (dataset, "container_doi"))
@@ -4117,6 +4133,10 @@ class ApiServer:
                                                        is_published = False)
 
                 if dataset is None:
+                    return self.error_403 (request)
+
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_edit", False)):
                     return self.error_403 (request)
 
                 is_embargoed = validator.boolean_value (record, "is_embargoed", when_none=False)
@@ -4186,6 +4206,10 @@ class ApiServer:
                                                            account_uuid=account_uuid,
                                                            is_published=False)
 
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_remove", False)):
+                    return self.error_403 (request)
+
                 container_uuid = dataset["container_uuid"]
                 if self.db.delete_dataset_draft (container_uuid, dataset["uuid"], account_uuid):
                     return self.respond_204()
@@ -4211,6 +4235,10 @@ class ApiServer:
                                                        account_uuid=account_uuid,
                                                        is_published=False)
 
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_read", False)):
+                    return self.error_403 (request)
+
                 authors = self.db.authors (item_uri   = dataset["uri"],
                                            account_uuid = account_uuid,
                                            is_published = False,
@@ -4229,6 +4257,17 @@ class ApiServer:
             parameters = request.get_json()
 
             try:
+                dataset = self.__dataset_by_id_or_uri (dataset_id,
+                                                       account_uuid=account_uuid,
+                                                       is_published=False)
+
+                if dataset is None:
+                    return self.error_403 (request)
+
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_edit", False)):
+                    return self.error_403 (request)
+
                 new_authors = []
                 records     = parameters["authors"]
                 for record in records:
@@ -4252,10 +4291,6 @@ class ApiServer:
                             self.log.error ("Adding a single author failed.")
                             return self.error_500()
                     new_authors.append(URIRef(uuid_to_uri (author_uuid, "author")))
-
-                dataset = self.__dataset_by_id_or_uri (dataset_id,
-                                                       account_uuid=account_uuid,
-                                                       is_published=False)
 
                 # The PUT method overwrites the existing authors, so we can
                 # keep an empty starting list. For POST we must retrieve the
@@ -4308,6 +4343,10 @@ class ApiServer:
             if dataset is None:
                 return self.error_403 (request)
 
+            if not (not value_or (dataset, "is_shared_with_me", False) or
+                    value_or (dataset, "metadata_edit", False)):
+                return self.error_403 (request)
+
             authors = self.db.authors (item_uri     = dataset["uri"],
                                        account_uuid = account_uuid,
                                        is_published = False,
@@ -4349,6 +4388,10 @@ class ApiServer:
                 if item is None:
                     return self.error_403 (request)
 
+                if not (not value_or (item, "is_shared_with_me", False) or
+                        value_or (item, "metadata_read", False)):
+                    return self.error_403 (request)
+
                 funding = self.db.fundings (item_uri     = item["uri"],
                                             account_uuid = account_uuid,
                                             is_published = False,
@@ -4372,6 +4415,10 @@ class ApiServer:
                                              is_published=False)
 
                 if item is None:
+                    return self.error_403 (request)
+
+                if not (not value_or (item, "is_shared_with_me", False) or
+                        value_or (item, "metadata_edit", False)):
                     return self.error_403 (request)
 
                 new_fundings = []
@@ -4452,6 +4499,10 @@ class ApiServer:
                                            is_published = False)
 
             if item is None:
+                return self.error_403 (request)
+
+            if not (not value_or (item, "is_shared_with_me", False) or
+                    value_or (item, "metadata_remove", False)):
                 return self.error_403 (request)
 
             fundings = self.db.fundings (item_uri     = item["uri"],
@@ -4594,6 +4645,10 @@ class ApiServer:
                                                        account_uuid = account_uuid,
                                                        is_published = False)
 
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_edit", False)):
+                    return self.error_403 (request)
+
                 # First, validate all values passed by the user.
                 # This way, we can be as certain as we can be that performing
                 # a PUT will not end in having no categories associated with
@@ -4661,6 +4716,10 @@ class ApiServer:
             if not dataset:
                 return self.response ("[]")
 
+            if not (not value_or (dataset, "is_shared_with_me", False) or
+                    value_or (dataset, "metadata_read", False)):
+                return self.error_403 (request)
+
             return self.response (json.dumps (formatter.format_dataset_embargo_record (dataset)))
 
         if request.method == 'DELETE':
@@ -4668,6 +4727,10 @@ class ApiServer:
                 dataset = self.__dataset_by_id_or_uri (dataset_id,
                                                        account_uuid = account_uuid,
                                                        is_published = False)
+
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "metadata_remove", False)):
+                    return self.error_403 (request)
 
                 if self.db.delete_dataset_embargo (dataset_uri = dataset["uri"],
                                                    account_uuid = account_uuid):
@@ -4693,6 +4756,10 @@ class ApiServer:
                                                              account_uuid=account_uuid,
                                                              is_published=False)
                 if dataset is None:
+                    return self.error_403 (request)
+
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "data_read", False)):
                     return self.error_403 (request)
 
                 files = self.db.dataset_files (
@@ -4755,6 +4822,10 @@ class ApiServer:
                 if dataset is None:
                     return self.error_403 (request)
 
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "data_edit", False)):
+                    return self.error_403 (request)
+
                 if link is not None:
                     file_id = self.db.insert_file (
                         dataset_uri        = dataset["uri"],
@@ -4811,6 +4882,10 @@ class ApiServer:
                 if dataset is None:
                     return self.error_404 (request)
 
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "data_read", False)):
+                    return self.error_403 (request)
+
                 metadata = self.__file_by_id_or_uri (file_id,
                                                      account_uuid = account_uuid,
                                                      dataset_uri = dataset["uri"])
@@ -4831,6 +4906,10 @@ class ApiServer:
                                                        is_published=False)
 
                 if dataset is None:
+                    return self.error_403 (request)
+
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "data_remove", False)):
                     return self.error_403 (request)
 
                 metadata = self.__file_by_id_or_uri (file_id,
@@ -4869,6 +4948,9 @@ class ApiServer:
             if dataset is None:
                 return self.error_404 (request)
 
+            if value_or (dataset, "is_shared_with_me", False):
+                return self.error_403 (request)
+
             links = self.db.private_links (item_uri   = dataset["uri"],
                                            account_uuid = account_uuid)
 
@@ -4882,6 +4964,9 @@ class ApiServer:
                                                             account_uuid = account_uuid)
                 if dataset is None:
                     return self.error_404 (request)
+
+                if value_or (dataset, "is_shared_with_me", False):
+                    return self.error_403 (request)
 
                 id_string = secrets.token_urlsafe()
                 expires_date = validator.date_value (parameters, "expires_date", False)
@@ -4944,6 +5029,10 @@ class ApiServer:
             return self.error_404 (request)
 
         if request.method in ("GET", "HEAD"):
+
+            if value_or (dataset, "is_shared_with_me", False):
+                return self.error_403 (request)
+
             links = self.db.private_links (
                         item_uri   = dataset["uri"],
                         id_string  = link_id,
@@ -4952,6 +5041,10 @@ class ApiServer:
             return self.default_list_response (links, formatter.format_private_links_record)
 
         if request.method == 'PUT':
+
+            if value_or (dataset, "is_shared_with_me", False):
+                return self.error_403 (request)
+
             parameters = request.get_json()
             try:
                 result = self.db.update_private_link (
@@ -4972,6 +5065,10 @@ class ApiServer:
                 return self.error_400 (request, error.message, error.code)
 
         if request.method == 'DELETE':
+
+            if value_or (dataset, "is_shared_with_me", False):
+                return self.error_403 (request)
+
             result = self.db.delete_private_links (dataset["container_uuid"],
                                                    account_uuid,
                                                    link_id)
@@ -5121,6 +5218,10 @@ class ApiServer:
                                                account_uuid = account_uuid)
 
         if dataset is None:
+            return self.error_403 (request)
+
+        if not (not value_or (dataset, "is_shared_with_me", False) or
+                value_or (dataset, "metadata_edit", False)):
             return self.error_403 (request)
 
         reserved_doi = self.__reserve_and_save_doi (account_uuid, dataset)
@@ -6079,7 +6180,7 @@ class ApiServer:
 
         return branch_name
 
-    def __git_repository_by_dataset_id (self, account_uuid, dataset_id):
+    def __git_repository_by_dataset_id (self, account_uuid, dataset_id, action="read"):
         """Deduplication for api_v3_datasets_git_[branches|files]."""
 
         dataset = self.__dataset_by_id_or_uri (dataset_id,
@@ -6088,6 +6189,10 @@ class ApiServer:
 
         if dataset is None:
             self.log.error ("No Git repository for dataset %s.", dataset_id)
+            return None
+
+        if not (not value_or (dataset, "is_shared_with_me", False) or
+                value_or (dataset, f"data_{action}", False)):
             return None
 
         # Pre-Djehuty datasets may not have a Git UUID. We therefore
@@ -6139,7 +6244,7 @@ class ApiServer:
         except validator.ValidationException as error:
             return self.error_400 (request, error.message, error.code)
 
-        git_repository = self.__git_repository_by_dataset_id (account_uuid, dataset_id)
+        git_repository = self.__git_repository_by_dataset_id (account_uuid, dataset_id, "edit")
         if git_repository is None:
             return self.error_404 (request)
 
@@ -6157,7 +6262,7 @@ class ApiServer:
         if account_uuid is None:
             return self.error_authorization_failed(request)
 
-        git_repository = self.__git_repository_by_dataset_id (account_uuid, dataset_id)
+        git_repository = self.__git_repository_by_dataset_id (account_uuid, dataset_id, "read")
         if git_repository is None:
             return self.error_404 (request)
 
@@ -6177,7 +6282,7 @@ class ApiServer:
         if account_uuid is None:
             return self.error_authorization_failed(request)
 
-        git_repository = self.__git_repository_by_dataset_id (account_uuid, dataset_id)
+        git_repository = self.__git_repository_by_dataset_id (account_uuid, dataset_id, "read")
         if git_repository is None:
             return self.error_404 (request)
 
@@ -6391,6 +6496,9 @@ class ApiServer:
         if dataset is None:
             self.locks.unlock (locks.LockTypes.SUBMIT_DATASET)
             return self.error_404 (request)
+
+        if value_or (dataset, "is_shared_with_me", False):
+            return self.error_403 (request)
 
         record = request.get_json()
         try:
@@ -6688,6 +6796,10 @@ class ApiServer:
             if dataset is None:
                 return self.error_403 (request)
 
+            if not (not value_or (dataset, "is_shared_with_me", False) or
+                    value_or (dataset, "data_edit", False)):
+                return self.error_403 (request)
+
             content_type = value_or (request.headers, "Content-Type", "")
             if not content_type.startswith ("multipart/form-data"):
                 return self.error_415 (["multipart/form-data"])
@@ -6940,6 +7052,10 @@ class ApiServer:
         if metadata is None:
             return self.error_404 (request)
 
+        if not (not value_or (metadata, "is_shared_with_me", False) or
+                value_or (metadata, "data_read", False)):
+            return self.error_403 (request)
+
         try:
             metadata["base_url"] = self.base_url
             return self.response (json.dumps (formatter.format_file_details_record (metadata)))
@@ -7017,6 +7133,10 @@ class ApiServer:
                                                account_uuid=account_uuid,
                                                is_published=False)
 
+        if not (not value_or (dataset, "is_shared_with_me", False) or
+                value_or (dataset, "metadata_read", False)):
+            return self.error_403 (request)
+
         return self.__api_v3_item_references (request, dataset)
 
     def api_v3_collection_references (self, request, collection_id):
@@ -7045,6 +7165,10 @@ class ApiServer:
             item  = item_by_id_procedure (item_id,
                                           account_uuid=account_uuid,
                                           is_published=False)
+
+            if not (not value_or (item, "is_shared_with_me", False) or
+                    value_or (item, "metadata_read", False)):
+                return self.error_403 (request)
 
             tags = self.db.tags (
                 item_uri        = item["uri"],
@@ -7250,6 +7374,9 @@ class ApiServer:
                 dataset = self.db.datasets (git_uuid = git_uuid, is_published=False)[0]
 
             if dataset is not None:
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "data_edit", False)):
+                    return self.error_403 (request)
                 return self.__git_passthrough (request)
         except IndexError:
             pass
@@ -7266,6 +7393,10 @@ class ApiServer:
                                             is_latest    = None)[0]
 
             if dataset is not None:
+                if not (not value_or (dataset, "is_shared_with_me", False) or
+                        value_or (dataset, "data_read", False)):
+                    return self.error_403 (request)
+
                 self.__log_event (request, dataset["container_uuid"], "dataset", "gitDownload")
                 return self.__git_passthrough (request)
         except IndexError:
